rename: networks -> chains
