+++ extension/src/transactionTranslations/types.ts
@@ -1,6 +1,6 @@
 import { MetaTransaction } from 'react-multisend'
 
-import { ChainId } from '../networks'
+import { ChainId } from '../chains'
 import { SupportedModuleType } from '../settings/Connection/useZodiacModules'
 
 export interface TransactionTranslation {
